socket logic for opening and editing tickets

+++ src/routes/TicketView.svelte
@@ -1,10 +1,24 @@
 <script>
     export let data;
 
-    import { outputDelay, renderMainView, renderTicketView, ROUTES, getData, sendRequest } from "./utils";
+    import { onMount, onDestroy } from 'svelte';
+    import { outputDelay, renderMainView, renderTicketView, ROUTES, getData, sendRequest, socket } from "./utils";
 
     const reasonCodes = getData(ROUTES.CODES);
     const existingTickets = getData(ROUTES.TICKETS);
+
+    let editedTickets = [];
+    const editedByClient = [];
+
+    onMount(() => {
+        // Receive updates
+        socket.on('ticket edit', function (_editedTickets) {
+            editedTickets = _editedTickets;
+        });
+
+        // Emit empty string to init editedTickets
+        socket.emit('ticket edit', "");
+    });
 
     // Create a new ticket and add it to db by posting to /tickets route
     async function createNewTicket() {
@@ -26,19 +40,14 @@
 
         await sendRequest(ROUTES.TICKETS, {_id: id, code: _code}, 'PUT');
 
-        renderTicketView(data);
-    }
-
-    // Delete a ticket
-    async function deleteTicket(id) {
-
-        await sendRequest(ROUTES.TICKETS, {_id: id}, 'DELETE');
+        removeClientEdit(id);
 
         renderTicketView(data);
     }
 
     // Open update tab
     async function openUpdateTicket(id) {
+        addClientEdit(id);
 
         let elem = document.getElementById(id);
         const oldCode = elem.innerText;
@@ -64,17 +73,59 @@
 
         // Listen for change
         selector.addEventListener('change', (event) => {
-            console.log(event.target.value);
             updateTicket(id, event.target.value)
         })
 
         elem.appendChild(selector);
     }
+
+    // Delete a ticket
+    async function deleteTicket(id) {
+
+    await sendRequest(ROUTES.TICKETS, {_id: id}, 'DELETE');
+
+    renderTicketView(data);
+    }
+
+    // Add an edit initiated by client
+    function addClientEdit(ticketId) {
+        editedTickets.push(ticketId);
+        editedByClient.push(ticketId);
+
+        socket.emit('ticket edit', editedTickets);
+    }
+
+    // Remove an edit initiated by client
+    function removeClientEdit(ticketId) {
+        editedTickets.splice(editedTickets.indexOf(ticketId), 1);
+        editedByClient.splice(editedByClient.indexOf(ticketId), 1);
+
+        socket.emit('ticket edit', editedTickets);
+    }
+
+    // Remove all edits initiated by client
+    function removeClientEdits() {
+        editedByClient.forEach((ticketId) => {
+            editedTickets.splice(editedTickets.indexOf(ticketId), 1);
+        });
+
+        socket.emit('ticket edit', editedTickets);
+    }
+
+    function goBack() {
+        removeClientEdits();
+
+        renderMainView();
+    }
+
+    onDestroy(() => {
+        removeClientEdits();
+    });
 </script>
 
 <div class="ticket-container">
     <!-- Back Button -->
-    <button on:click={renderMainView}>&larr; Tillbaka</button>
+    <button on:click={goBack}>&larr; Tillbaka</button>
 
     <h1>Nytt ärende</h1>
 
@@ -135,8 +186,13 @@
                 <td>{existingTicket.trainnumber}</td>
                 <td>{existingTicket.traindate}</td>
                 <td>
-                    <button on:click={() => openUpdateTicket(existingTicket._id)}>change</button>
-                    <button on:click={() => deleteTicket(existingTicket._id)}>del</button>
+                    {#if !editedTickets.includes(existingTicket._id)}
+                        <button on:click={() => openUpdateTicket(existingTicket._id)}>change</button>
+                        <button on:click={() => deleteTicket(existingTicket._id)}>del</button>
+                    {:else}
+                        <button disabled>change</button>
+                        <button disabled>del</button>
+                    {/if}
                 </td>
             </tr>
             {/each}
